requirements.txt
open_clip_torch==2.23.0 
transformers==4.35.2 
matplotlib
numpy 
scipy
scikit-learn
